CheckExpDesignCompleteness

This method check some experimental design characteristics.

does it have biological factor ?

does it have replicat/batch factor ?

does it have enough replicat/batch ? at least 3 are advised

is the design completed ? presence of all possible combinations of levels for all factors

is the design balanced ? presence of the same number of replicat for all possible combinations

Completed design and at least on biological and one batch factors are required for using RFLOMICS workflow.

# S4 method for ExpDesign
CheckExpDesignCompleteness(object, colnames = NULL)

Arguments

An
object of class ExpDesign-class

Value

a named list of two objects

"count:" a data.frame with the number of each possible combinations of levels for all factors.

"message:" results of the check

"true" :" false or "The experimental design is complete and balanced

"lowRep" :" false or "WARNING : 3 biological replicates are needed.

"noCompl" :" false or "ERROR : The experimental design is not complete.

"noBalan" :" warning or "WARNING : The experimental design is complete but not balanced.

"noBio" :" false or "ERROR : no bio factor !

"noBatch :" false or "ERROR : no replicat

Examples

Design.File <- read.table(file= paste(path.package("RFLOMICS"), "/ExamplesFiles/TP/experimental_design.txt",sep=""),header = TRUE,row.names = 1, sep = "\t")

# Define the type of each factor
Design.typeList <- c("Bio","Bio","batch")

# Define the reference modality for each factor
Design.refList <- c("WT","control","rep1")

# Initialize an object of class ExpDesign
Design.obj <- ExpDesign.constructor(ExpDesign = Design.File,
refList = Design.refList, typeList = Design.typeList)
CheckExpDesignCompleteness(Design.obj)
#> Error: Must group by variables found in `.data`.
#> * Column `temperature` is not found.
#> * Column `imbibition` is not found.