FlomicsMultiAssay.constructor Constructor for the class MultiAssayExperiment-class

This function initializes an object of class MultiAssayExperiment-class from a list of omics data and an object of class ExpDesign-class.

FlomicsMultiAssay.constructor(
  inputs,
  projectName,
  ExpDesign,
  refList,
  typeList
)

Arguments

inputs

A named list of omic dataset. Names must refer to the name of the omic dataset. An omics dataset must be itself a list of three objects:

data:matrix of omic data

meta:an optional quality check data

omicType:Type of omic data type "None", "RNAseq", "proteomics" or "Metabolomics".

projectName

Project name

ExpDesign

a data.frame. Row names give the name of each sample which has been to be construct by combining factor's modality separated by a "_" (EX: WT_treated_rep1). Column names give the name of an experimental factor which is a vector of character storing the factor modality for each sample.

refList

A list of string giving the reference modality for each factor.

typeList

A vector of string indicating the type of each experimental factor. Two types of effect are required ("Bio" or "batch")

Design

An object of class ExpDesign-class

Value

An object of class MultiAssayExperiment-class

Examples

Design.File <- read.table(file= paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/experimental_design.txt",sep=""),header = TRUE,row.names = 1, sep = "\t")

# Define the type of each factor
Design.Factors.Type <- c("Bio","Bio","batch")

# Define the reference modality for each factor
Design.Factors.Ref <- c("WT","control","rep1")

# Initialize an object of class ExpDesign
Design.obj <- ExpDesign.constructor(ExpDesign = Design.File, projectName = "Design.Name", refList = Design.Factors.Ref,
 typeList = Design.Factors.Type)
#> Error in ExpDesign.constructor(ExpDesign = Design.File, projectName = "Design.Name",     refList = Design.Factors.Ref, typeList = Design.Factors.Type): unused argument (projectName = "Design.Name")
Design.Factors.Name <- names(Design.File)
Design.formulae <- GetModelFormulae(Factors.Name = Design.Factors.Name,Factors.Type=Design.Factors.Type)
Design.formulae[[1]]
#> ~replicat + genotype + treatment + genotype:treatment
#> <environment: 0x7f88f2511ad8>
Design.obj <- getExpressionContrast(object = Design.obj, model.formula = names(Design.formulae[1]))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getExpressionContrast': object 'Design.obj' not found
Design.contrastList <- lapply(Design.obj@Contrasts.List, function(x) {
return(x[1:2]$contrast)
})
#> Error in lapply(Design.obj@Contrasts.List, function(x) {    return(x[1:2]$contrast)}): object 'Design.obj' not found
Design.obj <- getContrastMatrix(object = Design.obj, contrastList = unlist(Design.contrastList))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getContrastMatrix': object 'Design.obj' not found

 # Create a list of datasets
ListofData <- list("RNAseq1"=list("dataFile"=paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/rnaseq_gene_counts.txt",sep=""),
"qcFile"=paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/rnaseq_bioinfo_QC.txt",sep=""), "omicType"="RNAseq"))
FlomicsMultiAssay.constructor(inputs = ListofData, Design=Design.obj)
#> Error in FlomicsMultiAssay.constructor(inputs = ListofData, Design = Design.obj): unused argument (Design = Design.obj)