GetModelFormulae

From a vector of character giving the name of the factors of an omics experiment, and their type of effect: biological or batch, it returns all models formulae that can be formulated in association with this factors. Batch effect factors do not appear in interaction terms with biological factor. Model formulae stop in second order interaction.

GetModelFormulae(Factors.Name, Factors.Type)

Arguments

Factors.Name
a vector of character giving the name of the factors

Factors.Type
a vector of character giving the type of effect for the factor ("Bio" or "batch")

Value

a named list of object of class formula

Examples

GetModelFormulae(Factors.Name=c("Genotype","Temperature"),Factors.Type=c("Bio","Bio"))
#> $`~Genotype + Temperature + Genotype:Temperature`
#> ~Genotype + Temperature + Genotype:Temperature
#> <environment: 0x7fe742219b70>
#> 
#> $`~Genotype + Temperature`
#> ~Genotype + Temperature
#> <environment: 0x7fe73c803ee0>
#> 
#> $`~Temperature`
#> ~Temperature
#> <environment: 0x7fe73f64f6d8>
#> 
#> $`~Genotype`
#> ~Genotype
#> <environment: 0x7fe73f648e40>
#> 
GetModelFormulae(Factors.Name=c("Genotype","Temperature","Replicat"),Factors.Type=c("Bio","Bio","batch"))
#> $`~Replicat + Genotype + Temperature + Genotype:Temperature`
#> ~Replicat + Genotype + Temperature + Genotype:Temperature
#> <environment: 0x7fe742320e18>
#> 
#> $`~Replicat + Genotype + Temperature`
#> ~Replicat + Genotype + Temperature
#> <environment: 0x7fe73ff58350>
#> 
#> $`~Replicat + Temperature`
#> ~Replicat + Temperature
#> <environment: 0x7fe7422f8e88>
#> 
#> $`~Replicat + Genotype`
#> ~Replicat + Genotype
#> <environment: 0x7fe7422ea440>
#> 
GetModelFormulae(Factors.Name=c("Genotype","Temperature","Environment"),Factors.Type=c("Bio","Bio","Bio"))
#> $`~Genotype + Temperature + Environment + Genotype:Temperature + Genotype:Environment + Temperature:Environment`
#> ~Genotype + Temperature + Environment + Genotype:Temperature + 
#>     Genotype:Environment + Temperature:Environment
#> <environment: 0x7fe7428a4190>
#> 
#> $`~Temperature + Environment + Temperature:Environment`
#> ~Temperature + Environment + Temperature:Environment
#> <environment: 0x7fe742842190>
#> 
#> $`~Genotype + Temperature + Genotype:Temperature`
#> ~Genotype + Temperature + Genotype:Temperature
#> <environment: 0x7fe742827888>
#> 
#> $`~Genotype + Environment + Genotype:Environment`
#> ~Genotype + Environment + Genotype:Environment
#> <environment: 0x7fe742831a48>
#> 
#> $`~Genotype + Temperature + Environment`
#> ~Genotype + Temperature + Environment
#> <environment: 0x7fe7428b5748>
#> 
#> $`~Temperature + Environment`
#> ~Temperature + Environment
#> <environment: 0x7fe74287b660>
#> 
#> $`~Genotype + Temperature`
#> ~Genotype + Temperature
#> <environment: 0x7fe742863b68>
#> 
#> $`~Genotype + Environment`
#> ~Genotype + Environment
#> <environment: 0x7fe74286e668>
#> 
#> $`~Temperature`
#> ~Temperature
#> <environment: 0x7fe73ffe3e08>
#> 
#> $`~Environment`
#> ~Environment
#> <environment: 0x7fe73ffed158>
#> 
#> $`~Genotype`
#> ~Genotype
#> <environment: 0x7fe73ffd1f58>
#> 
GetModelFormulae(Factors.Name=c("Genotype","Temperature","Environment","Replicat"),Factors.Type=c("Bio","Bio","Bio","batch"))
#> $`~Replicat + Genotype + Temperature + Environment + Genotype:Temperature + Genotype:Environment + Temperature:Environment`
#> ~Replicat + Genotype + Temperature + Environment + Genotype:Temperature + 
#>     Genotype:Environment + Temperature:Environment
#> <environment: 0x7fe743060b70>
#> 
#> $`~Replicat + Temperature + Environment + Temperature:Environment`
#> ~Replicat + Temperature + Environment + Temperature:Environment
#> <environment: 0x7fe74302a780>
#> 
#> $`~Replicat + Genotype + Temperature + Genotype:Temperature`
#> ~Replicat + Genotype + Temperature + Genotype:Temperature
#> <environment: 0x7fe743005ea8>
#> 
#> $`~Replicat + Genotype + Environment + Genotype:Environment`
#> ~Replicat + Genotype + Environment + Genotype:Environment
#> <environment: 0x7fe743021fc8>
#> 
#> $`~Replicat + Genotype + Temperature + Environment`
#> ~Replicat + Genotype + Temperature + Environment
#> <environment: 0x7fe74306e4e0>
#> 
#> $`~Replicat + Temperature + Environment`
#> ~Replicat + Temperature + Environment
#> <environment: 0x7fe743050470>
#> 
#> $`~Replicat + Genotype + Temperature`
#> ~Replicat + Genotype + Temperature
#> <environment: 0x7fe74303cfd0>
#> 
#> $`~Replicat + Genotype + Environment`
#> ~Replicat + Genotype + Environment
#> <environment: 0x7fe743047660>
#> 
#> $`~Replicat + Temperature`
#> ~Replicat + Temperature
#> <environment: 0x7fe73efe6558>
#> 
#> $`~Replicat + Environment`
#> ~Replicat + Environment
#> <environment: 0x7fe73eff5e80>
#> 
#> $`~Replicat + Genotype`
#> ~Replicat + Genotype
#> <environment: 0x7fe73efdf978>
#>