RflomicsSE-class Class

RflomicsSE is a class that extends the SummarizedExperiment by imposing a structure on the metadata slot. This class is used by the Rflomics analysis workflow to store the experimental design, the settings and results of a single omic analysis. The slot metadata is structured as follows:

# S4 method for RflomicsSE
getDesignMat(object)

# S4 method for RflomicsSE
getDatasetNames(object)

# S4 method for RflomicsSE
getOmicsTypes(object)

# S4 method for RflomicsSE
getFactorNames(object)

# S4 method for RflomicsSE
getFactorTypes(object)

# S4 method for RflomicsSE
getBioFactors(object)

# S4 method for RflomicsSE
getBatchFactors(object)

# S4 method for RflomicsSE
getMetaFactors(object)

# S4 method for RflomicsSE
getFactorModalities(object, factorName)

Arguments

object

An object of class RflomicsSE

factorName

factor name

Value

A RflomicsSE object.

Slots

See SummarizedExperiment

The slot metadata is structured as follows:

omicType: the type of omics dataset

design: experimental design

DataProcessing: a list containing the data processing settings and results

PCAlist: a list containing the PCA settings and results

DiffExpAnal: a list containing the Differential Analysis settings and results

CoExpAnal: a list containing the Coexpression Analysis settings and results

DiffExpEnrichAnal: a list containing the enrichment analysis of the list of DE features settings and results

CoExpEnrichAnal: a list containing the enrichment analysis of the list of co-expressed features settings and results

Accessors

getDesignMat: return a data.frame with experimental design.

getDatasetNames: return a string with dataset name.

getOmicsTypes: return a named vector with omics type of dataset ("RNAseq", "proteomics", "metabolomics")

getFactorNames: return a vector with the experimental factor names.

getFactorTypes: return a named vector with experimental factor types ("bio", "batch" or "meta").

getBioFactors: return a vector with the biological factor names.

getBatchFactors: return a vector with the batch factor names.

getMetaFactors: return a vector with the metadata factor names.

getFactorModalities: return a vector with the modality names of selected factor.

See also