RunDiffAnalysis

# S4 method for SummarizedExperiment
RunDiffAnalysis(
  object,
  design,
  Adj.pvalue.method = "BH",
  contrastList,
  DiffAnalysisMethod,
  clustermq = FALSE
)

Arguments

object

an object of class [SummarizedExperiment]} {Adj.pvalue.method}{The method choosen to adjust pvalue.} {contrastList}{The list of contrast to test} {DiffAnalysisMethod}{A character vector giving the name of the differential analysis method to run. Either "edgeRglmfit", "limmalmFit", ...} {clustermq}{A boolean indicating whether the constrasts have to be computed in local or in a distant machine} {Adj.pvalue.cutoff}{The adjusted pvalue cut-off} {filter_only}{A boolean indicating whether only filter on DE results have to be applied (filter_only=TRUE). FALSE by default.} } { All the results are stored as a named list DiffExpAnal in the metadata slot of a given SummarizedExperiment object. Objects are:

contrasts: The selected contrasts for which the differential analysis has been conducted

method: The method used for the differential analysis

Adj.pvalue.method: The method applied for the pvalue adjustment

Adj.pvalue.cutoff: The threshold applied for the pvalue adjustment

FDR: The false discovery rate given in input

RawDEFres: a list giving for each contrast the raw results of the differential analysis method

DEF: a list giving for each contrast a data.frame of non filtered differential expressed features

TopDEF: a list giving for each contrast a data.frame of differential expressed features by Adj.pvalue.cutoff

mergeDEF: A data frame indicating for each features in row, if it is DE in a given contrasts in column

An object of class [SummarizedExperiment]

This is an interface method which run a differential analysis method on omic datasets stored in an object of class SummarizedExperiment. According to the type of omic and to a list of contrasts, a differential analysis method is applied to each contrasts (or hypothesis). Three methods are available according to the type of object:

For RNAseq data: the glmFit function of the edgeR package

For proteomic and metabolomic data: the lmFit function of the limma package

Parameters used for RNAseq are those recommended in DiCoExpress workflow (see the paper in reference)
Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool to process multifactorial RNAseq experiments from quality controls to co-expression analysis through differential analysis based on contrasts inside GLM models. Plant Methods 16, 68 (2020).