Feature selection in a Rflomics MAE

This function selects all the features to keep according to user's choices on each omic data.

# S4 method for RflomicsMAE
filterFeatures(
  object,
  selOpt = rep(list("all"), length(object)),
  type = rep(list("union"), length(selOpt))
)

Arguments

object

An object of class RflomicsMAE-class. It is expected the MAE object is produced by rflomics previous analyses, as it relies on their results.

selOpt

list of vectors. Preferred named list with names corresponding to the names of the experimentList in the object. For each Experiment, gives the option for the filtering: either 'all', 'DE', 'none', or a specific name of a contrast or cluster (if coexpression results are available). Default is taking all features for all experiment list. If the vector is named and an Experiment is missing, no feature will be selected from it. If the vector is not named, the selection will be applied in order of the Experiments in the object.

type

if selOpt is set to a specific set of contrasts or clusters, indicates whether the selection is "union" or "intersection" of entities in these sets.

Value

a RflomicsMAE, filtered with only the corresponding features.

Examples

# load ecoseed data
library(RFLOMICS)
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- createRflomicsMAE(
    projectName = "Tests",
    omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
    omicsNames  = c("RNAtest", "metatest", "protetest"),
    omicsTypes  = c("RNAseq","metabolomics","proteomics"),
    ExpDesign   = ecoseed$design,
    factorRef   = ecoseed$factorRef)
names(MAE) <- c("RNAtest", "metatest", "protetest")

formulae <- generateModelFormulae( MAE) 
MAE <- setModelFormula(MAE, formulae[[1]])
contrastList <- Reduce(rbind, generateExpressionContrast(MAE)) 

MAE <- MAE |>
    setSelectedContrasts(contrastList[c(3,6,25)]) |>
    runTransformData(SE.name = "metatest", transformMethod = "log2") |>
    runNormalization(SE.name = "metatest", normMethod = "median")    |>
    runNormalization(SE.name = "protetest", normMethod = "median")   |>
    runDiffAnalysis(SE.name = "metatest", method = "limmalmFit")     |>
    runDiffAnalysis(SE.name = "protetest", method = "limmalmFit")    |>
    runCoExpression(SE.name = "protetest", K = 2:10, replicates = 5, 
                    merge = "union")

# Define the selections options
selOpt = list("protetest" = c("cluster.1", "H3"), "metatest" = c("DE"))
MAE1 <- filterFeatures(MAE, selOpt)
#> harmonizing input:
#>   removing 27 sampleMap rows not in names(experiments)
MAE1
#> A RflomicsMAE object of 2 listed
#>  experiments with user-defined names and respective classes.
#>  Containing an ExperimentList class object of length 2:
#>  [1] protetest: RflomicsSE with 30 rows and 27 columns
#>  [2] metatest: RflomicsSE with 47 rows and 27 columns
#> Functionality:
#>  experiments() - obtain the ExperimentList instance
#>  colData() - the primary/phenotype DataFrame
#>  sampleMap() - the sample coordination DataFrame
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment
#>  *Format() - convert into a long or wide DataFrame
#>  assays() - convert ExperimentList to a SimpleList of matrices
#>  exportClass() - save data to flat files

selOpt2 = list("protetest" = c("cluster.2", "H25"), metatest = c("DE"))
MAE2 <- filterFeatures(MAE, selOpt2,
                       type = c("metatest" = "intersection",
                                "protetest" = 'union'))
#> harmonizing input:
#>   removing 27 sampleMap rows not in names(experiments)
MAE2
#> A RflomicsMAE object of 2 listed
#>  experiments with user-defined names and respective classes.
#>  Containing an ExperimentList class object of length 2:
#>  [1] protetest: RflomicsSE with 25 rows and 27 columns
#>  [2] metatest: RflomicsSE with 47 rows and 27 columns
#> Functionality:
#>  experiments() - obtain the ExperimentList instance
#>  colData() - the primary/phenotype DataFrame
#>  sampleMap() - the sample coordination DataFrame
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment
#>  *Format() - convert into a long or wide DataFrame
#>  assays() - convert ExperimentList to a SimpleList of matrices
#>  exportClass() - save data to flat files