Generate expression of contrasts

Generate expression of contrasts based on chosed model formula.

generateExpressionContrast: This function allows, from a model formulae, to give the expression contrast data frames. Three types of contrasts are expressed:

pairwise comparison

averaged expression

interaction expression

setSelectedContrasts: Set the selected contrasts stored in metadata slot

getSelectedContrasts: List the selected contrasts

# S4 method for RflomicsSE
generateExpressionContrast(object, contrastType = NULL)

# S4 method for RflomicsMAE
generateExpressionContrast(object, contrastType = NULL)

# S4 method for RflomicsMAE
setSelectedContrasts(object, contrastList = NULL)

# S4 method for RflomicsSE
setSelectedContrasts(object, contrastList = NULL)

# S4 method for RflomicsMAE
getSelectedContrasts(object)

# S4 method for RflomicsSE
getSelectedContrasts(object)

Arguments

object

an object of class RflomicsSE or class RflomicsMAE-class

contrastType

type of contrasts from which the possible contrasts are extracted ("average", "simple", "interaction"). Default is all contrasts types.

contrastList

a data.frame of contrasts generated by generateExpressionContrast

Value

list of 1 or 3 data.frames of contrast expression

an object of RflomicsSE class or RflomicsMAE-class

class

Author

Christine Paysant-Le Roux, adapted by Nadia Bessoltane

Examples

# load ecoseed data
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- RFLOMICS::createRflomicsMAE(
  projectName = "Tests",
  omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
  omicsNames  = c("RNAtest", "metatest", "protetest"),
  omicsTypes  = c("RNAseq","metabolomics","proteomics"),
  ExpDesign   = ecoseed$design,
  factorRef   = ecoseed$factorRef)

# generate all statistical model formulae
formulae <- generateModelFormulae(MAE)

# chose and set model formula to rflomicsMAE object
MAE <- setModelFormula(MAE, formulae[[1]])

# Generate expression of contrasts from chosen model
contrastList <- generateExpressionContrast(MAE, "averaged")

# Set the contrasts List and choose the first 3 contrasts of type averaged
MAE <- setSelectedContrasts(MAE, contrastList = contrastList[c(1,2,3),])