Preparation step for integration

This function transforms a RflomicsMAE produced by rflomics into an untrained MOFA object or a list to use for mixOmics. It checks for batch effects to correct them before integration. It also transforms RNASeq counts data into continuous data using voom. This is the second step into the integration. It is usually preceded by filterFeatures to extract the correct variables, and followed by runOmicsIntegration.

# S4 method for RflomicsMAE
prepareForIntegration(
  object,
  omicsNames = NULL,
  rnaSeq_transfo = "limma (voom)",
  variableLists = NULL,
  group = NULL,
  method = "MOFA",
  transformData = TRUE,
  cmd = FALSE
)

Arguments

object

An object of class RflomicsMAE-class. It is expected the MAE object is produced by rflomics previous analyses, as it relies on their results.

omicsNames

vector of characters strings, referring to the names of the filtered table in 'object@ExperimentList'.

rnaSeq_transfo

character string, only supports 'limma (voom)' for now. Transformation of the rnaSeq data from counts to continuous data.

variableLists

list of variables to keep per dataset.

group

Not implemented yet in the interface. Useful for MOFA2 run.

method

one of MOFA or mixOmics. Method for which the object is prepared.

transformData

boolean. Transform the data with the transform and normalization method? Default is TRUE.

cmd

used in the interface. Print cmd lines.

Value

An untrained MOFA object or a list of dataset

Examples

# load ecoseed data
library(RFLOMICS)
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- createRflomicsMAE(
    projectName = "Tests",
    omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
    omicsNames  = c("RNAtest", "metatest", "protetest"),
    omicsTypes  = c("RNAseq","metabolomics","proteomics"),
    ExpDesign   = ecoseed$design,
    factorRef   = ecoseed$factorRef)
names(MAE) <- c("RNAtest", "metatest", "protetest")

formulae <- generateModelFormulae( MAE) 
MAE <- setModelFormula(MAE, formulae[[1]])

contrastList <- generateExpressionContrast(object = MAE) |> 
    purrr::reduce(rbind) |>
    dplyr::filter(contrast %in% c("(temperatureElevated_imbibitionDS - temperatureLow_imbibitionDS)",
                                  "((temperatureLow_imbibitionEI - temperatureLow_imbibitionDS) + (temperatureMedium_imbibitionEI - temperatureMedium_imbibitionDS) + (temperatureElevated_imbibitionEI - temperatureElevated_imbibitionDS))/3",
                                  "((temperatureElevated_imbibitionEI - temperatureLow_imbibitionEI) - (temperatureElevated_imbibitionDS - temperatureLow_imbibitionDS))" ))
MAE <- MAE |>
    setSelectedContrasts(contrastList) |>
    runTransformData(SE.name = "metatest", transformMethod = "log2") |>
    runNormalization(SE.name = "metatest", normMethod = "median")    |>
    runNormalization(SE.name = "protetest", normMethod = "median")   |>
    runDiffAnalysis(SE.name = "metatest", method = "limmalmFit")     |>
    runDiffAnalysis(SE.name = "protetest", method = "limmalmFit")    

# Integration using MOFA
# Prepare mofa object:
mofaObj <- prepareForIntegration(MAE,
                                 omicsNames = c("protetest", "metatest"),
                                 variableLists = rownames(MAE),
                                 method = "MOFA")
#> Warning: 'experiments' dropped; see 'drops()'
#> harmonizing input:
#>   removing 27 sampleMap rows not in names(experiments)
class(mofaObj)
#> [1] "MOFA"
#> attr(,"package")
#> [1] "MOFA2"

# Integration using MixOmics
mixObj <- prepareForIntegration(MAE,
                                omicsNames = c("protetest", "metatest"),
                                variableLists = rownames(MAE),
                                method = "mixOmics")
#> Warning: 'experiments' dropped; see 'drops()'
#> harmonizing input:
#>   removing 27 sampleMap rows not in names(experiments)
class(mixObj)
#> [1] "list"