Run Gene Enrichment Analysis and process results

This function performs overrepresentation analysis (ORA) using clusterprofiler functions. It can be used with custom annotation file (via enricher), GO (enrichGO) or KEGG (enrichKEGG) annotations.

# S4 method for RflomicsSE
runAnnotationEnrichment(
  object,
  nameList = NULL,
  list_args = list(pvalueCutoff = 0.05, qvalueCutoff = 1, minGSSize = 10, maxGSSize =
    500, universe = names(object)),
  from = "DiffExp",
  database = "custom",
  domain = "no-domain",
  col_term = "term",
  col_gene = "gene",
  col_name = "name",
  col_domain = NULL,
  annot = NULL
)

# S4 method for RflomicsMAE
runAnnotationEnrichment(
  object,
  SE.name,
  nameList = NULL,
  list_args = list(pvalueCutoff = 0.05, qvalueCutoff = 1, minGSSize = 10, maxGSSize =
    500, universe = names(object)),
  from = "DiffExp",
  database = "custom",
  domain = "no-domain",
  col_term = "term",
  col_gene = "gene",
  col_name = "name",
  col_domain = NULL,
  annot = NULL
)

# S4 method for RflomicsSE
plotKEGG(
  object,
  contrastName,
  pathway_id = NULL,
  species = "ath",
  gene_idtype = "kegg",
  from = "DiffExp",
  ...
)

# S4 method for RflomicsSE
plotClusterProfiler(
  object,
  contrastName,
  database,
  domain = "no-domain",
  from = "DiffExp",
  plotType = "dotplot",
  showCategory = 15,
  searchExpr = "",
  nodeLabel = "all",
  p.adj.cutoff = NULL,
  ...
)

# S4 method for RflomicsSE
plotEnrichComp(
  object,
  from = "DiffExp",
  database = NULL,
  domain = "no-domain",
  matrixType = "FC",
  nClust = NULL,
  ...
)

# S4 method for RflomicsSE
getEnrichRes(
  object,
  contrastName = NULL,
  from = "DiffExpEnrichAnal",
  database = "GO",
  domain = NULL,
  ...
)

# S4 method for RflomicsMAE
getEnrichRes(
  object,
  experiment,
  contrastName = NULL,
  from = "DiffExpEnrichAnal",
  database = "GO",
  domain = NULL,
  ...
)

# S4 method for RflomicsSE
sumORA(
  object,
  from = "DiffExpEnrichAnal",
  database = NULL,
  contrastName = NULL
)

# S4 method for RflomicsSE
getEnrichPvalue(object, from = "DiffExpEnrichAnal", database = "GO")

# S4 method for RflomicsSE
getEnrichSettings(object, from = "DiffExpEnrichAnal", database = "GO")

# S4 method for RflomicsMAE
getAnnotAnalysesSummary(
  object,
  from = "DiffExpEnrichAnal",
  matrixType = "presence",
  ...
)

Arguments

object

An object of class RflomicsSE or class RflomicsMAE-class

nameList

name of contrasts (tags or names) from which to extract DE genes if from is DiffExpAnal.

list_args

list of arguments to pass to the enrichment function. These arguments must match the ones from the clusterprofiler package. E.g: universe, keytype, pvalueCutoff, qvalueCutoff, etc.

from

indicates if the enrichment results are taken from differential analysis results (DiffExpEnrichAnal) or from the co-expression analysis results (CoExpEnrichAnal)

database

which database (GO, KEGG, custom...)

domain

the subonology or subdomain for the database (eg CC, MF or BP for GO.)

col_term

for custom annotation, column name of the file containing term names or id.

col_gene

for custom annotation, column name of the file containing entities names.

col_name

(optional) for custom annotation, column name of the file containing term names if col_term is used for ids.

col_domain

for custom annotation, column name of the file containing the domains.

annot

for custom annotation, path of the annotation file.

SE.name

SE.name the name of the dataset if the input object is a RflomicsMAE-class

contrastName

the contrastName or clusterName to retrieve the results from. If NULL, all results are returned.

pathway_id

the KEGG id pathway to plot.

species

Kegg code (eg hsa or ath)

gene_idtype

idtype (kegg, uniprot,...)

...

more arguments for ComplexHeatmap::Heatmap.

plotType

type of plot. Define the function used inside. One of dotplot, heatplot or cnetplot.

showCategory

max number of terms to show.

searchExpr

expression to search in the showCategory terms.

nodeLabel

same as in enrichplot::cnetplot function, defines

p.adj.cutoff

pvalueCutoff to define the enrichment threshold.

matrixType

Heatmap matrix to plot, one of GeneRatio, p.adjust or presence.

nClust

number of separate cluster to plot on the heatmap, based o n the clustering.

experiment

if the object is a RflomicsMAE, then experiment is the name of the RflomicsSE to look for.

Value

A RflomicsMAE or a RflomicsS, depending on the class of object parameter. The enrichment results are added to the metadata slot, either in DiffExpEnrichAnal or CoExpEnrichAnal.

Only displays the KEGG pathway, it does not return any object.

a list of tables or a table

the pvalue cutoff used for the analysis.

a list with all settings

Accessors

A set of getters and setters generic functions to access and modify objects of the slot metadata of a RflomicsMAE-class object or a RflomicsMAE-class object.

getEnrichRes: get a particular enrichment result. return enrichment results given in the form of lists of clusterprofiler results.

sumORA: Get summary tables from ORA analyses - once an enrichment has been conducted.

getEnrichPvalue: Get the pvalue threshold used in enrichment analysis.

getEnrichSettings: get the settings of an enrichment analysis.

getAnnotAnalysesSummary: return A list of heatmaps, one for each ontology/domain.

Plots

A collection of functions for plotting results from omic analysis steps.

plotKEGG: Plot the KEGG pathway using the pathview package. It overlays the result of the differential analysis or the clusters entity on the image.

plotClusterProfiler: Plot a dotplot, a cnetplot or an heatplot, using enrichplot package. It is a wrapper method destined for the RflomicsSE class..

plotEnrichComp: plot an heatmap of all the enriched term found for a given database and a given source (differential analysis or coexpression clusters). Allow for the comparison of several enrichment results.

Examples

# load ecoseed data
library(RFLOMICS)
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- createRflomicsMAE(
    projectName = "Tests",
    omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
    omicsNames  = c("RNAtest", "metatest", "protetest"),
    omicsTypes  = c("RNAseq","metabolomics","proteomics"),
    ExpDesign   = ecoseed$design,
    factorRef   = ecoseed$factorRef)
# names(MAE) <- c("RNAtest", "metatest", "protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- generateExpressionContrast(MAE, "averaged")

MAE <- setSelectedContrasts(MAE, contrastList = contrastList[c(1, 2, 3),])

# Run the data preprocessing and perform the differential analysis 
MAE <- runDataProcessing(MAE, SE.name = "protetest",  
                         transformMethod = "log2",
                         normMethod = "median")
#> [RFLOMICS] #    => select samples...
#> [RFLOMICS] #    => transformation data...
#> [RFLOMICS] #    => Run normalization...
#> [RFLOMICS] #    => Compute PCA 

MAE <- runDiffAnalysis(MAE, 
                       SE.name = "protetest", 
                       method = "limmalmFit")

# Run GO annotation (enrichGO)
# Not run: need org.At.tair.db package
# MAE <- runAnnotationEnrichment(MAE, SE.name = "protetest",
#                                list_args = list(OrgDb = "org.At.tair.db",
#                                                 keyType = "TAIR",
#                                                 pvalueCutoff = 0.05),
#                                from = "DiffExp", database = "GO",
#                                domain = "CC")

# Run KEGG annotation (enrichKEGG)
# need internet connection
MAE <- runAnnotationEnrichment(MAE, SE.name = "protetest",
                               list_args = list(organism = "ath",
                                                keyType = "kegg",
                                                pvalueCutoff = 0.05),
                               from = "DiffExp", database = "KEGG")
#> warning: 

# Search for the pvalue cutoff:
sumORA(MAE[["protetest"]], from = "DiffExp", database = "KEGG")
#>                                                                                            Contrast
#> (temperatureMedium - temperatureLow) in mean           (temperatureMedium - temperatureLow) in mean
#> (temperatureElevated - temperatureLow) in mean       (temperatureElevated - temperatureLow) in mean
#> (temperatureElevated - temperatureMedium) in mean (temperatureElevated - temperatureMedium) in mean
#>                                                   no-domain
#> (temperatureMedium - temperatureLow) in mean              5
#> (temperatureElevated - temperatureLow) in mean           31
#> (temperatureElevated - temperatureMedium) in mean        23

# need internet connection
# pathview package
# plotKEGG(MAE[["protetest"]], pathway_id = "ath00710", species = "ath",
#          contrastName = "cluster.4", from = "Coexp")

# From differential analysis proteins lists:
# plotClusterProfiler(MAE[["protetest"]],
#                     contrastName = "(temperatureElevated - temperatureMedium) in mean",
#                     database = "KEGG", from = "DiffExp",
#                     plotType = "heatplot", p.adj.cutoff = 0.05,
#                     domain = "no-domain")
# 
# plotEnrichComp(MAE[["protetest"]], from = "DiffExp",
#                database = "KEGG", matrixType = "FC")

# Get all results from KEGG on differential expression lists:
results <- getEnrichRes(MAE[["protetest"]],
                        from = "diffexp", database = "KEGG")

# Search for the pvalue cutoff:
usedPvalue <- 
    getEnrichPvalue(MAE[["protetest"]], from = "diffexp", database = "KEGG")
settings <- 
    getEnrichSettings(MAE[["protetest"]], from = "diffexp", database = "KEGG")