runCoExpression

This method performs a co-expression/co-abundance analysis of omic-data.

# S4 method for RflomicsSE
runCoExpression(
  object,
  K = 2:20,
  replicates = 5,
  contrastNames = NULL,
  merge = "union",
  model = "Normal",
  GaussianModel = NULL,
  transformation = NULL,
  normFactors = NULL,
  clustermq = FALSE,
  meanFilterCutoff = NULL,
  scale = NULL,
  silent = TRUE,
  cmd = FALSE
)

# S4 method for RflomicsMAE
runCoExpression(
  object,
  SE.name,
  K = 2:20,
  replicates = 5,
  contrastNames,
  merge = "union",
  model = "Normal",
  GaussianModel = NULL,
  transformation = NULL,
  normFactors = NULL,
  clustermq = FALSE,
  meanFilterCutoff = NULL,
  scale = NULL,
  silent = TRUE,
  cmd = FALSE
)

# S4 method for RflomicsMAE
getCoExpAnalysesSummary(object, omicNames = NULL, ...)

# S4 method for RflomicsSE
plotCoExpression(object)

# S4 method for RflomicsMAE
plotCoExpression(object, SE.name)

# S4 method for RflomicsSE
plotCoExpressionProfile(
  object,
  cluster = 1,
  condition = "groups",
  features = NULL
)

# S4 method for RflomicsMAE
plotCoExpressionProfile(
  object,
  SE.name,
  cluster = 1,
  condition = "groups",
  features = NULL
)

# S4 method for RflomicsSE
plotCoseqContrasts(object)

# S4 method for RflomicsMAE
plotCoseqContrasts(object, SE.name)

# S4 method for RflomicsSE
getCoexpSettings(object)

# S4 method for RflomicsMAE
getCoexpSettings(object, SE.name)

# S4 method for RflomicsSE
getClusterEntities(object, clusterName)

# S4 method for RflomicsMAE
getClusterEntities(object, SE.name, clusterName)

# S4 method for RflomicsSE
getCoseqClusters(object)

# S4 method for RflomicsMAE
getCoseqClusters(object, SE.name)

Arguments

object

An object of class RflomicsSE or class RflomicsMAE-class

K

Number of clusters (a single value or a vector of values)

replicates

The number of iteration for each K.

contrastNames

names of the contrasts from which the DE entities have to be taken. Can be NULL, in that case every contrasts from the differential analysis are taken into consideration.

merge

"union" or "intersection"

model

Type of mixture model to use "Poisson" or "normal". By default, it is the normal.

GaussianModel

Type of GaussianModel to be used for the Normal mixture model only. This parameters is set to "Gaussian_pk_Lk_Ck" by default and doesn't have to be changed except if an error message proposed to try another model like "Gaussian_pk_Lk_Bk".

transformation

The transformation type to be used. By default, it is the "arcsin" one.

normFactors

The type of estimator to be used to normalize for differences in library size. By default, it is the "TMM" one.

clustermq

boolean. Does the computation need to be executed on a distant server. A configuration file and a network connection are needed to use this option.

meanFilterCutoff

a cutoff to filter a gene with a mean expression lower than this value. (only for RNAseq data, set to NULL for others).

scale

Boolean. If TRUE scale all variables tounit variance.

silent

if TRUE, coseq run silently (without any console print or message)

cmd

if TRUE, print steps of the analysis. Used inside the coseq module in the shiny interface.

SE.name

SE.name the name of the dataset if the input object is a RflomicsMAE-class

omicNames

the name of the experiment to summarize.

...

Not in use at the moment

cluster

cluster number

condition

Default is group.

features

Default is NULL.

clusterName

name of the cluster

Value

An S4 object of class RflomicsSE

All the results are stored as a named list CoExpAnal

in the metadata slot of a given RflomicsSE object. Objects are: The runCoExpression method return several results, for coseq

method, objects are:

setting: co-expression analysis settings. See getCoexpSetting

results: boolean indicating if the co-expression analysis succeed

coseqResults: the raw results of coseq

clusters: a List of clusters

cluster.nb: The number of cluster

plots: The plots of coseq results

stats: A tibble summarising failed jobs: reason, propoif any

Details

For now, only the coseq function of the coseq package is used. For RNAseq data, parameters used are those recommended in DiCoExpress workflow (see the reference). This parameters are: model="normal", transformation="arcsin", GaussianModel="Gaussian_pk_Lk_Ck", normFactors="TMM", meanFilterCutoff = 50 For proteomic or metabolomic, data are scaled by protein or metabolite to group them by expression profiles rather than by expression intensity. After data scaling, recommended parameters (from coseq developers) for co-expression analysis are: model="normal", transformation="none", GaussianModel="Gaussian_pk_Lk_Ck", normFactors="none", meanFilterCutoff = NULL.

Accessors

A set of getters and setters generic functions to access and modify objects of the slot metadata of a RflomicsMAE-class object or a RflomicsMAE-class object.

getCoexpSettings: Access to the co-expression analysis settings of a given omic dataset

getClusterEntities: get members of a cluster. return The list of entities inside this cluster.

getCoseqClusters: get cluster. Return all clusters

Plots

A collection of functions for plotting results from omic analysis steps.

getCoExpAnalysesSummary: ...

plotCoExpression: list plot of ICL, logLike and coseq object with min ICL

plotCoExpressionProfile: ...

plotCoseqContrasts: This function describes the composition of clusters according to the contrast to which the gene belongs

References

Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool to process multifactorial RNAseq experiments from quality controls to co-expression analysis through differential analysis based on contrasts inside GLM models. Plant Methods 16, 68 (2020).

See also

Examples

# load ecoseed data
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- RFLOMICS::createRflomicsMAE(
  projectName = "Tests",
  omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
  omicsNames  = c("RNAtest", "metatest", "protetest"),
  omicsTypes  = c("RNAseq","metabolomics","proteomics"),
  ExpDesign   = ecoseed$design,
  factorRef   = ecoseed$factorRef)

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- generateExpressionContrast(MAE, "averaged")

MAE <- setSelectedContrasts(MAE, contrastList = contrastList[c(1, 2, 3),])

# Run the data preprocessing and perform the differential analysis 
MAE <- runDataProcessing(MAE, SE.name = "protetest",  
                         transformMethod = "log2",
                         normMethod = "median")
#> [RFLOMICS] #    => select samples...
#> [RFLOMICS] #    => transformation data...
#> [RFLOMICS] #    => Run normalization...
#> [RFLOMICS] #    => Compute PCA 
MAE <- runDiffAnalysis(MAE, SE.name = "protetest")
#> Warning: DiffAnalyseMethod was missing. Detected omic type is proteomics using limmalmFit for differential analysis.

# Run co-expression analysis
MAE <- runCoExpression(MAE, SE.name = "protetest", 
                       K = 2:5, replicates = 5, 
                       merge = "union")

# get parametres used to run co-expression analysis
coExp.set.list <- getCoexpSettings(MAE[["protetest"]])
coExp.set.list$method
#> [1] "coseq"

# get results
clusters <- getCoseqClusters(MAE[["protetest"]])
proteins_c1 <- getClusterEntities(MAE[["protetest"]], 
                                      clusterName = "cluster_1")

# plots
#plotCoExpression(MAE[["protetest"]])

#plotCoExpressionProfile(MAE[["protetest"]], cluster = 2)

#plotCoseqContrasts(MAE[["protetest"]])