runCoExpression

This is an interface method which performed co-expression/co-abundance analysis of omic-data.

# S4 method for SummarizedExperiment
runCoExpression(
  object,
  geneList,
  K = 2:20,
  replicates = 5,
  nameList,
  merge = "union",
  model = "Normal",
  GaussianModel = "Gaussian_pk_Lk_Ck",
  transformation,
  normFactors,
  clustermq = FALSE
)

Arguments

object
An object of class SummarizedExperiment

geneList
A list of genes

K
Number of clusters (a single value or a vector of values)

replicates
The number of iteration for each K.

merge
"union" or "intersection"

model
Type of mixture model to use "Poisson" or "normal". By default, it is the normal.

GaussianModel
Type of GaussianModel to be used for the Normal mixture model only. This parameters is set to "Gaussian_pk_Lk_Ck" by default and doesn't have to be changed except if an error message proposed to try another model like "Gaussian_pk_Lk_Bk".

transformation
The transformation type to be used. By default, it is the "arcsin" one.

normFactors
The type of estimator to be used to normalize for differences in library size. By default, it is the "TMM" one.

Value

All the results are stored as a named list CoExpAnal in the metadata slot of a given SummarizedExperiment object. Objects are: The runCoExpression method return several results, for coseq method, objects are:

model: see model params description

transformation: see transformation params description

normFactors: see normFactors params description

meanFilterCutoff: set to 50 for RNA and to NULL for others

gene.list.names: see nameList in Arguments description

merge.type: see merge params description

coseqResults: the raw results of coseq

clusters: a List of clusters

cluster.nb: The number of cluster

plots: The plots of coseq results

An S4 object of class SummarizedExperiment

Details

References

Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool to process multifactorial RNAseq experiments from quality controls to co-expression analysis through differential analysis based on contrasts inside GLM models. Plant Methods 16, 68 (2020).

See also