Data Exploratory and processing

These functions applied a data processing (filtering, normalization and/or transformation, PCA) on RNAseq, proteomic, or metabolomic data.

runDataProcessing() calls the following functions:

filterLowAbundance: This function aims at removing transcript from the count data matrix of an omic of type "RNAseq". by applying filtering criterion described in reference.

runSampleFiltering: This function applied sample filtering on an dataset.

runTransformData: This function applied a transformation on dataset. The transformation method is chosen according to dataset omic type (RNaseq: none, metabolomics/proteomics: log2)

runNormalization: This function applied a normalization on a dataset. The normalisation method is chosen according to dataset omic type (RNAseq: TMM, metabolics/proteomics: median)

runOmicsPCA: This function performs a principal component analysis on omic data stored in an object of class RflomicsSE-class Results are stored in the metadata slot of the same object. If a "Normalization" slot is present in the metadata slot, then data are normalized before running the PCA according to the indicated transform method.

This function performs a principal component analysis on omic data stored in an object of class RflomicsSE-class Results are stored in the metadata slot of the same object. If a "Normalization" slot is present in the metadata slot, then data are normalized before running the PCA according to the indicated transform method.

checkExpDesignCompleteness: return a string with message. This method checks some experimental design characteristics. A complete design (all combinations of factor modalities with at least 2 replicates for each have to be present) with at least one biological and one batch factors are required to use the RFLOMICS workflow.

plotDataDistribution: return boxplot or density plot of expression or abundance distribution.

# S4 method for RflomicsSE
runDataProcessing(
  object,
  samples = NULL,
  lowCountFiltering_strategy = "NbReplicates",
  lowCountFiltering_CPM_Cutoff = 1,
  normMethod = "none",
  transformMethod = "none"
)

# S4 method for RflomicsMAE
runDataProcessing(
  object,
  samples = NULL,
  lowCountFiltering_strategy = "NbReplicates",
  lowCountFiltering_CPM_Cutoff = 1,
  normMethod = "none",
  transformMethod = "none",
  SE.name
)

# S4 method for RflomicsSE
filterLowAbundance(
  object,
  filterMethod = "CPM",
  filterStrategy = "NbConditions",
  cpmCutoff = 5
)

# S4 method for RflomicsMAE
filterLowAbundance(
  object,
  SE.name,
  filterMethod = "CPM",
  filterStrategy = "NbConditions",
  cpmCutoff = 5
)

# S4 method for RflomicsSE
runSampleFiltering(object, samples = NULL)

# S4 method for RflomicsMAE
runSampleFiltering(object, SE.name, samples = NULL)

# S4 method for RflomicsSE
runTransformData(object, transformMethod = NULL, modifyAssay = FALSE)

# S4 method for RflomicsMAE
runTransformData(object, SE.name, transformMethod = NULL, modifyAssay = FALSE)

# S4 method for RflomicsSE
runNormalization(object, normMethod = NULL, modifyAssay = FALSE)

# S4 method for RflomicsMAE
runNormalization(object, SE.name, normMethod = NULL, modifyAssay = FALSE)

# S4 method for RflomicsSE
runOmicsPCA(object, ncomp = 5, raw = FALSE)

# S4 method for RflomicsMAE
runOmicsPCA(object, SE.name, ncomp = 5, raw = FALSE)

# S4 method for RflomicsSE
checkExpDesignCompleteness(object, sampleList = NULL)

# S4 method for RflomicsMAE
checkExpDesignCompleteness(object, omicName, sampleList = NULL)

# S4 method for RflomicsSE
getProcessedData(object, NormTrans = TRUE)

# S4 method for RflomicsMAE
getProcessedData(object, SE.name, NormTrans = TRUE)

# S4 method for RflomicsSE
getTransSettings(object)

# S4 method for RflomicsMAE
getTransSettings(object, SE.name)

# S4 method for RflomicsSE
getFilterSettings(object)

# S4 method for RflomicsMAE
getFilterSettings(object, SE.name)

# S4 method for RflomicsSE
getFilteredFeatures(object)

# S4 method for RflomicsMAE
getFilteredFeatures(object, SE.name)

# S4 method for RflomicsSE
getCoeffNorm(object)

# S4 method for RflomicsMAE
getCoeffNorm(object, SE.name)

# S4 method for RflomicsSE
getNormSettings(object)

# S4 method for RflomicsMAE
getNormSettings(object, SE.name)

# S4 method for RflomicsSE
plotLibrarySize(object, raw = FALSE)

# S4 method for RflomicsMAE
plotLibrarySize(object, SE.name, raw = FALSE)

# S4 method for RflomicsSE
plotDataDistribution(object, plot = "boxplot", raw = FALSE)

# S4 method for RflomicsMAE
plotDataDistribution(object, SE.name, plot = "boxplot", raw = FALSE)

# S4 method for RflomicsSE
plotOmicsPCA(
  object,
  raw = c("raw", "norm"),
  axes = c(1, 2),
  groupColor = "groups"
)

# S4 method for RflomicsMAE
plotOmicsPCA(object, SE.name, raw, axes = c(1, 2), groupColor = "groups")

# S4 method for RflomicsSE
plotExpDesignCompleteness(object, sampleList = NULL)

# S4 method for RflomicsMAE
plotExpDesignCompleteness(object, omicName, sampleList = NULL)

Arguments

object

An object of class RflomicsSE-class.

samples

samples to keep.

lowCountFiltering_strategy

strategy of RNAseq low count filtering. Mandatory for RNAseq data. Default value: "NbReplicates".

lowCountFiltering_CPM_Cutoff

CPM cutoff for RNAseq low count filtering. Mandatory for RNAseq data. Default value: 1.

normMethod

Normalization method

transformMethod

The transformation to store in the metadata or to store and apply if modifyAssay is TRUE.

SE.name

the name of the data the normalization have to be applied to.

filterMethod

The filtering model ("CPM")

filterStrategy

The filtering strategy ("NbConditions" or "NbReplicates")

cpmCutoff

The CPM cutoff.

modifyAssay

Does the normalization have to be applied or just stored for later? Recommended it stays FALSE.

ncomp

Number of components to compute. Default is 5.

raw

This argument indicates whether the scaled PCA has to be performed on raw [‘raw’] or normalized [‘norm’] data.

sampleList

list of samples to check.

omicName

a character string with the name of the dataset

NormTrans

boolean. If TRUE, returned Data are normalized and transformed.

plot

plot type ("boxplot" or "density")

axes

A vector giving the two axis that have to be drawn for the factorial map

groupColor

All combination of level's factor

Value

An object of class RflomicsSE or class RflomicsSE

An object of class RflomicsSE

The applied normalization method and computed scaling factors (by samples) are stored as a named list ("normalization") of two elements (respectively "method" and "coefNorm") in the metadata slot of a given data set, stored itself in the ExperimentList slot of a RflomicsSE object.

An object of class RflomicsSE

Details

filterLowAbundance(): By default, gene/transcript with 0 count are removed from the data. The function then computes the count per million or read (CPM) for each gene in each sample and gives by genes the number of sample(s) which are over the cpmCutoff (NbOfsample_over_cpm). Then Two filtering strategies are proposed:

NbConditions: keep gene if the NbOfsample_over_cpm >= NbConditions

NbReplicates: keep gene if the NbOfsample_over_cpm >= min(NbReplicat)

filterByExpr: the default filtering method implemented in the edgeR filterByExpr() function.

Accessors

getProcessedData: return a processed data matrix (filtering, normalization and/or transformation)

getTransSettings: return a list of transformation settings of a given omic dataset

getFilterSettings: return a list the filtering settings of a given omic dataset

getFilteredFeatures: return a vector of filtered features of a given omic dataset

getCoeffNorm: return a named vector with normalization coefficients of a given omic dataset

getNormSettings: return a list of normalization settings of a given omic dataset

Plots

plotLibrarySize: return barplot of library size by sample.

plotOmicsPCA: This function plot the factorial map from a PCA object stored in a RflomicsSE-class object. By default, samples are colored by groups (all combinations of level's factor)

plotExpDesignCompleteness: This method checks that experimental design constraints are satisfied and plot a summary of the design. A complete design (all combinations of factor modalities with at least 2 replicates for each have to be present) with at least one biological and one batch factors are required to use the RFLOMICS workflow.

References

Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool to process multifactorial RNAseq experiments from quality controls to co-expression analysis through differential analysis based on contrasts inside GLM models. Plant Methods 16, 68 (2020).

See also

Examples

# load ecoseed data
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- RFLOMICS::createRflomicsMAE(
  projectName = "Tests",
  omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
  omicsNames  = c("RNAtest", "metatest", "protetest"),
  omicsTypes  = c("RNAseq","metabolomics","proteomics"),
  ExpDesign   = ecoseed$design,
  factorRef   = ecoseed$factorRef)

# Set the statistical model
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# set the contrast list
contrastList <- generateExpressionContrast(MAE, "averaged")
MAE <- setSelectedContrasts(MAE, contrastList = contrastList[c(1, 2, 3),])

# Data processing of RNAseq dataset : RNAtest
## using data processing functions for RNAseq data
### filter low RNAseq count 
# MAE <- filterLowAbundance(MAE, SE.name = "RNAtest",
#                           filterStrategy = "NbReplicates", 
#                           cpmCutoff = 1)
# ### filter outlier samples
# MAE <- runSampleFiltering(MAE, SE.name = "RNAtest", 
#                           samples = colnames(MAE[["RNAtest"]])[-1])
# ### data normalisation outlier samples
# MAE <- runNormalization(MAE, SE.name = "RNAtest", 
#                         normMethod = "TMM")

## use runDataProcessing function that combines the previous three functions
MAE <- runDataProcessing(MAE, SE.name = "RNAtest", 
                         samples = colnames(MAE[["RNAtest"]])[-1], 
                         lowCountFiltering_strategy = "NbReplicates", 
                         lowCountFiltering_CPM_Cutoff = 1, 
                         normMethod = "TMM") 
#> [RFLOMICS] #    => select samples...
#> [RFLOMICS] #    => Low counts Filtering...
#> [RFLOMICS] #    => Counts normalization...
#> [RFLOMICS] #    => Compute PCA 

## check completness of RNAtest data
checkExpDesignCompleteness(MAE, omicName = "RNAtest")$messages
#> [1] "The experimental design is complete and balanced."

# Data processing of proteimics dataset : protetest
# ## transform data
# MAE <- runTransformData(MAE, SE.name = "protetest",  transformMethod = "log2")
# ## normalise data
# MAE <- runNormalization(MAE, SE.name = "protetest", normMethod = "median") 

## use runDataProcessing function
MAE <- runDataProcessing(MAE, SE.name = "protetest", 
                         normMethod = "median", 
                         transformMethod = "log2")
#> [RFLOMICS] #    => select samples...
#> [RFLOMICS] #    => transformation data...
#> [RFLOMICS] #    => Run normalization...
#> [RFLOMICS] #    => Compute PCA 

# plotExpDesignCompleteness(MAE[["RNAtest"]])

# plot Library Size
# plotDataDistribution(MAE[["RNAtest"]], raw=TRUE)
# plotDataDistribution(MAE[["RNAtest"]], raw=FALSE)

# plot gene expression distribution
# plotDataDistribution(MAE[["RNAtest"]], raw=TRUE, plot = "boxplot")
# plotDataDistribution(MAE[["RNAtest"]], raw=FALSE, plot = "boxplot")

# plot PCA 
# plotOmicsPCA(MAE[["RNAtest"]], raw="raw", groupColor = "imbibition")
# plotOmicsPCA(MAE[["RNAtest"]], raw="norm", groupColor = "imbibition")

# See runDataProcessing for an example that includes getProcessedData
# See runDataProcessing for an example that includes getTransSettings
# See runDataProcessing for an example that includes getFilterSettings
# See runDataProcessing for an example that includes getFilteredFeatures
# See runDataProcessing for an example that includes getCoeffNorm
# See runDataProcessing for an example that includes getNormSettings
# See runDataProcessing for an example that includes plotLibrarySize
# See runDataProcessing for an example that includes plotDataDistribution
# See runDataProcessing for an example that includes plotOmicsPCA
# See runDataProcessing for an example that includes plotExpDesignCompleteness