Run Differential Expression Analysis and process results

This is an interface method which run a differential analysis on omics datasets stored in an object of class RflomicsSE or RflomicsMAE-class. According to the type of omics and to a list of contrasts, a differential analysis is performef for each contrasts. Two methods are available according to the type of object:

For RNAseq data: the glmFit function/model of the edgeR package is applied.

For proteomic and metabolomic data: the lmFit function/model of the limma package is applied.

generateContrastMatrix: Defines contrast matrix or contrast list with contrast name and contrast coefficients

filterDiffAnalysis: The filterDiffAnalysis method allows filtering the results of the differential analysis based on a new cutoff for p-value and fold change.

setValidContrasts: Set the valid contrasts stored in metadata slot.

# S4 method for RflomicsSE
runDiffAnalysis(
  object,
  contrastList = NULL,
  method = NULL,
  p.adj.method = "BH",
  p.adj.cutoff = 0.05,
  logFC.cutoff = 0,
  clustermq = FALSE,
  cmd = FALSE,
  ...
)

# S4 method for RflomicsMAE
runDiffAnalysis(
  object,
  SE.name,
  contrastList = NULL,
  method = NULL,
  p.adj.method = "BH",
  p.adj.cutoff = 0.05,
  logFC.cutoff = 0,
  clustermq = FALSE,
  cmd = FALSE,
  ...
)

# S4 method for RflomicsSE
generateContrastMatrix(object, contrastList = NULL)

# S4 method for RflomicsMAE
generateContrastMatrix(object, SE.name, contrastList = NULL)

# S4 method for RflomicsSE
filterDiffAnalysis(object, p.adj.cutoff = 0.05, logFC.cutoff = 0)

# S4 method for RflomicsMAE
filterDiffAnalysis(object, SE.name, p.adj.cutoff = NULL, logFC.cutoff = NULL)

# S4 method for RflomicsSE
setValidContrasts(object, contrastList = NULL)

# S4 method for RflomicsMAE
setValidContrasts(object, omicName = NULL, contrastList = NULL)

# S4 method for RflomicsSE
plotDiffAnalysis(
  object,
  contrastName,
  typeofplots = c("MA.plot", "volcano", "histogram")
)

# S4 method for RflomicsMAE
plotDiffAnalysis(
  object,
  SE.name,
  contrastName,
  typeofplots = c("MA.plot", "volcano", "histogram")
)

# S4 method for RflomicsSE
plotHeatmapDesign(
  object,
  contrastName,
  splitFactor = "none",
  title = "",
  annotNames = NULL,
  modalities = NULL,
  drawArgs = list(),
  heatmapArgs = list()
)

# S4 method for RflomicsMAE
plotHeatmapDesign(
  object,
  SE.name,
  contrastName,
  splitFactor = "none",
  title = "",
  annotNames = NULL,
  modalities = NULL,
  drawArgs = list(),
  heatmapArgs = list()
)

# S4 method for RflomicsSE
plotBoxplotDE(object, features = NULL, groupColor = "groups", raw = FALSE)

# S4 method for RflomicsMAE
plotBoxplotDE(
  object,
  SE.name,
  features = NULL,
  groupColor = "groups",
  raw = FALSE
)

# S4 method for RflomicsSE
getDEMatrix(object)

# S4 method for RflomicsMAE
getDEMatrix(object, SE.name)

# S4 method for RflomicsSE
getDEList(object, contrasts = NULL, operation = "union")

# S4 method for RflomicsMAE
getDEList(object, SE.name, contrasts = NULL, operation = "union")

# S4 method for RflomicsSE
getDiffSettings(object)

# S4 method for RflomicsMAE
getDiffSettings(object, SE.name)

# S4 method for RflomicsSE
getValidContrasts(object)

# S4 method for RflomicsMAE
getValidContrasts(object, omicName)

# S4 method for RflomicsSE
sumDiffExp(object)

# S4 method for RflomicsMAE
sumDiffExp(object, SE.name = NULL)

# S4 method for RflomicsMAE
getDiffAnalysesSummary(
  object,
  plot = FALSE,
  ylabelLength = 30,
  nbMaxLabel = 20,
  interface = FALSE
)

Arguments

object

An object of class RflomicsSE or class RflomicsMAE-class

contrastList

A data.frame of contrast

method

A character vector giving the name of the differential analysis method to run. Either "edgeRglmfit" or "limmalmFit".

p.adj.method

The method choosen to adjust pvalue. Takes the same values as the ones of adj.p.adjust method.

p.adj.cutoff

adjusted pvalue cutoff. Default is the parameter from the differential analysis.

logFC.cutoff

cutoff for absolute value of log2FC. Default is the parameter from the differential analysis.

clustermq

A boolean indicating whether the constrasts have to be computed in local or in a distant machine

cmd

boolean. Used in the interface. If TRUE, print cmd for the user.

...

Additional arguments.

SE.name

SE.name the name of the dataset if the input object is a RflomicsMAE-class

omicName

a dataset name

contrastName

The contrastName for which the MAplot has to be drawn

typeofplots

The plots you want to return. Default is all possible plots: MA plot, Volcano plot and non adjusted pvalues histogram.

splitFactor

characters. Default to none. Name of a feature in the design matrix, splits the samples on the heatmap according to its modalities.

title

characters. Title of the heatmap.

annotNames

vector. Names of the annotations to keep in the Heatmap. Default takes all available information.

modalities

named list of vectors of modalities to subset and print on the heatmap.

drawArgs, heatmapArgs

named lists. Any additional parameter passed to ComplexHeatmap::Heatmap or ComplexHeatmap::draw

features

variable name (gene/protein/metabolite name)

groupColor

default to groups, indicate a variable in the design to color the boxplots accordingly.

raw

Boolean. Plot the raw data or the transformed ones (TRUE)

contrasts

Vector of characters, expect to be contrast names. Default is null, the operation (union) is performed on every contrasts found.

operation

character. Either union or intersection. Defines the operation to perform on the DE lists from the contrasts.

plot

FALSE or TRUE

ylabelLength

max length of the labels (characters)

nbMaxLabel

number of labels to print

interface

boolean. Is this plot for the interace or commandline?

Value

A RflomicsSE or a RflomicsMAE-class object. All the results are stored as a named list DiffExpAnal

in the metadata slot of a given RflomicsSE object. Objects are:

stats: data.frame giving a summary of the differential statistical analysis results by contrast: number of DE features, number of up and down regulated features

setting: Parameters used for the differential analysis

method: The method used for the differential analysis

p.adj.method: The applied p-value correction method

p.adj.cutoff: The cut-off applied for the adjusted p-value

logFC.cutoff: The absolute log FC cut-off

RawDEFres: a list giving for each contrast the raw results of the differential analysis method

DEF: a list giving for each contrast a data.frame of non filtered differential expressed features with their statistics

TopDEF: a list giving for each contrast a data.frame of differential expressed features ordered and filtered by p.adj.cutoff with their statistics

mergeDEF: a data frame of 0 and 1 indicating for each features in row, if it is DE in a given contrasts in column

contrasts: a data.table of the contrasts used for the differential analysis

a data.frame with differential analyses summary

Details

Functions and parameters used for RNAseq are those recommended in DiCoExpress workflow (see the paper in reference). Functions and parameters used for proteomic and metabolomic data are those recommended in the (Efstathiou *et al.*, 2017)

Accessors

getDEMatrix: return a matrix of experimental design.

getDEList: return a vector of union or intersection of differential expressed features from list of contrasts.

getDiffSettings: retrun a list of differential expression analysis settings of a given omic dataset

getValidContrasts: return a data.frame of valided contrasts

sumDiffExp: Get summary table from diffExpAnalysis analysis

getDiffAnalysesSummary: ...

Plots

plotDiffAnalysis method draws a MAplot, a volcano plot and the p-values distribution from the results of a differential analysis.

plotHeatmapDesign method draws a heatmap from the results of a differential analysis.

plotBoxplotDE method draws a boxplot showing the expression of given differentially expressed feature.

References

Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool to process multifactorial RNAseq experiments from quality controls to co-expression analysis through differential analysis based on contrasts inside GLM models. Plant Methods 16, 68 (2020).

Efstathiou G, Antonakis AN, Pavlopoulos GA, et al. ProteoSign: an end-user online differential proteomics statistical analysis platform. Nucleic Acids Res. 2017;45(W1):W300-W306.

See also

Author

Christine Paysant-Le Roux, adapted by Nadia Bessoltane

Examples

# load ecoseed data
library(RFLOMICS)
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- createRflomicsMAE(
  projectName = "Tests",
  omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
  omicsNames  = c("RNAtest", "metatest", "protetest"),
  omicsTypes  = c("RNAseq","metabolomics","proteomics"),
  ExpDesign   = ecoseed$design,
  factorRef   = ecoseed$factorRef)

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- generateExpressionContrast(MAE, "averaged")

MAE <- setSelectedContrasts(MAE, contrastList = contrastList[c(1, 2, 3),])

# Run the data preprocessing and perform the differential analysis 
MAE <- runDataProcessing(MAE, SE.name = "protetest",  
                         transformMethod = "log2",
                         normMethod = "median")
#> [RFLOMICS] #    => select samples...
#> [RFLOMICS] #    => transformation data...
#> [RFLOMICS] #    => Run normalization...
#> [RFLOMICS] #    => Compute PCA 

MAE <- runDiffAnalysis(MAE, SE.name = "protetest", 
                       method = "limmalmFit")
# or
# MAE[["protetest"]] <- runDiffAnalysis(MAE[["protetest"]], 
#                                       method = "limmalmFit", 
#                                       contrastList = contrastList)

# Filter the results of the differential analysis with new cut-off values 
# for p-value and fold change.
MAE <- filterDiffAnalysis(MAE, SE.name = "protetest",
                          p.adj.cutoff = 0.01,
                          logFC.cutoff = 0)
# or
# MAE[["protetest"]] <- filterDiffAnalysis(MAE[["protetest"]], 
#                                          p.adj.cutoff = 0.01, 
#                                          logFC.cutoff = 0)

# Access to the diff analysis settings
## Get DE matrix from DiffExpAnalysis
head(getDEMatrix(MAE[["protetest"]]))
#>         DEF (temperatureMedium - temperatureLow) in mean
#> 1 AT2G42560                                            1
#> 2 AT2G33070                                            1
#> 3 AT4G27140                                            1
#> 4 AT5G54740                                            1
#> 5 AT5G58070                                            1
#> 6 AT1G14940                                            1
#>   (temperatureElevated - temperatureLow) in mean
#> 1                                              1
#> 2                                              1
#> 3                                              1
#> 4                                              1
#> 5                                              1
#> 6                                              1
#>   (temperatureElevated - temperatureMedium) in mean
#> 1                                                 0
#> 2                                                 0
#> 3                                                 0
#> 4                                                 0
#> 5                                                 1
#> 6                                                 0

## Get union or intersection from list of contrasts
getDEList(MAE[["protetest"]], contrasts = "(temperatureMedium - temperatureLow) in mean") 
#>  [1] "AT2G42560" "AT2G33070" "AT4G27140" "AT5G54740" "AT5G58070" "AT1G14940"
#>  [7] "AT1G77120" "AT4G04460" "AT1G20620" "AT2G38380" "AT3G21380" "AT5G57550"
#> [13] "AT2G21660" "AT1G14950" "AT1G70840" "AT3G60730" "AT1G17100" "AT4G15530"
#> [19] "AT3G54400" "AT5G52640" "AT2G28490" "AT2G23110" "AT2G24270" "AT1G14930"
#> [25] "AT2G46750" "AT5G66190" "AT3G14067" "AT3G12203" "AT3G09440" "AT5G26120"
#> [31] "AT3G44300" "AT5G12950" "AT1G65880" "AT2G41260" "AT5G35360" "AT1G43670"
#> [37] "AT5G55730" "AT4G34480" "AT1G18500" "AT3G23490" "AT5G43600" "AT5G34940"
#> [43] "AT5G04885" "AT1G52560" "AT4G19810" "AT2G43090" "AT1G20020" "AT1G26560"
#> [49] "AT4G35790" "AT3G07770" "AT4G16760"

## Get diff setting
getDiffSettings(MAE[["protetest"]])
#> $method
#> [1] "limmalmFit"
#> 
#> $p.adj.method
#> [1] "BH"
#> 
#> $p.adj.cutoff
#> [1] 0.01
#> 
#> $abs.logFC.cutoff
#> [1] 0
#> 

# generate plot results of a differential analysis
thiscontrast <- "(temperatureMedium - temperatureLow) in mean"

## generate MAplot from diff analysis
# plotDiffAnalysis(MAE[["protetest"]], 
#                  contrastName = thiscontrast, 
#                  typeofplots = "MA.plot")

## plot the heatmap
# plotHeatmapDesign(MAE[["protetest"]], 
#                   contrastName = thiscontrast)

## plot boxplot with feature expression
# plotBoxplotDE(MAE[["protetest"]], 
#               features = "AT1G47128", 
#               groupColor = "temperature")
# plotBoxplotDE(MAE[["protetest"]], 
#               features = "AT1G79550", 
#               groupColor = "imbibition")

# See runDiffAnalysis for an example that includes plotDiffAnalysis
# See runDiffAnalysis for an example that includes plotHeatmapDesign
# See runDiffAnalysis for an example that includes plotBoxplotDE
# See runDiffAnalysis for an example that includes getDEMatrix
# See runDiffAnalysis for an example that includes getDEList
# See runDiffAnalysis for an example that includes getDiffSettings