runOmicsIntegration

Runs the integration according to the selected method (MOFA or mixOmics) and the settings given by the user. Requires to have the correct entry format in preparedObject before running.

These methods are used to directly access the results of multi-omics analyses or their settings, usually stored in the metadata of the RflomicsMAE-class object. Setters are also available.

# S4 method for RflomicsMAE
runOmicsIntegration(
  object,
  preparedObject = NULL,
  method = "MOFA",
  scale_views = FALSE,
  maxiter = 1000,
  num_factors = 10,
  selectedResponse = NULL,
  ncomp = 2,
  link_datasets = 1,
  link_response = 1,
  sparsity = FALSE,
  cases_to_try = 5,
  cmd = FALSE,
  ...
)

# S4 method for RflomicsMAE
getMixOmics(object, response = NULL, onlyResults = TRUE)

# S4 method for RflomicsMAE
getMOFA(object, onlyResults = TRUE)

# S4 method for RflomicsMAE
getMOFASettings(object)

# S4 method for RflomicsMAE
getMixOmicsSettings(object)

# S4 method for RflomicsMAE
setMOFA(object, results = NULL)

# S4 method for RflomicsMAE
setMixOmics(object, results = NULL)

# S4 method for RflomicsMAE
sumMixOmics(object, selectedResponse = NULL)

Arguments

object

a MAE object (produced by Flomics).

preparedObject

An untrained MOFA object or a list of dataset. Usually a result of prepareForIntegration.

method

one of MOFA or mixOmics. Method for which the object is prepared.

scale_views

boolean. If TRUE, scale each dataset to unit variance.

maxiter

MOFA2 parameter. Number of max iteration (otherwise stop when converged.)

num_factors

MOFA2 parameter. The number of factor to compute.

selectedResponse

a character. Useful if MixOmics was run on several response variable. If NULL, all variables are taken into account.

ncomp

mixOmics parameter. Number of components to compute.

link_datasets

mixOmics parameter. Link between datasets in the computation.

link_response

mixOmics parameter. Link between dataset and response.

sparsity

boolean. Used to determine which mixOmics function to apply (either block.plsda if FALSE or block.splsda if TRUE).

cases_to_try

integer. If sparsity is set to TRUE, then cases_to_try is used to determine the number of sets of variables to test for tuning.

cmd

boolean. Used in the interface. If TRUE, print cmd in the console.

...

not in use at the moment

response

a character giving the response variable to access specifically.

onlyResults

default return only the MixOmics or MOFA2 results. If you want to access all information of the integration, set onlyResults to FALSE. In MixOmics case, works only when response is specified.

results

The MOFA or mixOmics results to set in the object. If null, set to NULL.

Value

a RflomicsMAE object with the correct metadata slot filled with the results and the settings.

For getters: in getMixOmics, if response is NULL, then all the mixOmics results are returned. Otherwise, it gives the particular mixOmics result. For MOFA, returns the untrained object and the trained object as a list.

For setters: always returns a RflomicsMAE-class object.

sumMixOmics: A data frame or a list of dataframe (if selectedResponse is NULL) presenting the summary of mixOmics analyses.

Examples

# load ecoseed data
library(RFLOMICS)
data(ecoseed)

# create rflomicsMAE object with ecoseed data
MAE <- createRflomicsMAE(
    projectName = "Tests",
    omicsData   = list(ecoseed$RNAtest, ecoseed$metatest, ecoseed$protetest),
    omicsNames  = c("RNAtest", "metatest", "protetest"),
    omicsTypes  = c("RNAseq","metabolomics","proteomics"),
    ExpDesign   = ecoseed$design,
    factorRef   = ecoseed$factorRef)
names(MAE) <- c("RNAtest", "metatest", "protetest")

formulae <- generateModelFormulae( MAE) 
MAE <- setModelFormula(MAE, formulae[[1]])

contrastList <- generateExpressionContrast(object = MAE) |> 
    purrr::reduce(rbind) |>
    dplyr::filter(contrast %in% c("(temperatureElevated_imbibitionDS - temperatureLow_imbibitionDS)",
                                  "((temperatureLow_imbibitionEI - temperatureLow_imbibitionDS) + (temperatureMedium_imbibitionEI - temperatureMedium_imbibitionDS) + (temperatureElevated_imbibitionEI - temperatureElevated_imbibitionDS))/3",
                                  "((temperatureElevated_imbibitionEI - temperatureLow_imbibitionEI) - (temperatureElevated_imbibitionDS - temperatureLow_imbibitionDS))" ))
MAE <- MAE |>
    setSelectedContrasts(contrastList) |>
    runTransformData(SE.name = "metatest", transformMethod = "log2") |>
    runNormalization(SE.name = "metatest", normMethod = "median")    |>
    runNormalization(SE.name = "protetest", normMethod = "median")   |>
    runDiffAnalysis(SE.name = "metatest", method = "limmalmFit")     |>
    runDiffAnalysis(SE.name = "protetest", method = "limmalmFit")    

# Integration using MOFA
# Prepare mofa object:
mofaObj <- prepareForIntegration(MAE,
                                 omicsNames = c("protetest", "metatest"),
                                 variableLists = rownames(MAE),
                                 method = "MOFA")
#> Warning: 'experiments' dropped; see 'drops()'
#> harmonizing input:
#>   removing 27 sampleMap rows not in names(experiments)

# Perform integration:
MAEtest <- runOmicsIntegration(MAE, 
                               preparedObject = mofaObj, 
                               method = "MOFA")
#> Checking data options...
#> Checking training options...
#> Checking model options...
#> Warning: The total number of samples is very small for learning 10 factors.  
#>     Try to reduce the number of factors to obtain meaningful results. It should not exceed ~6.
#> Connecting to the mofapy2 python package using reticulate (use_basilisk = FALSE)... 
#>     Please make sure to manually specify the right python binary when loading R with reticulate::use_python(..., force=TRUE) or the right conda environment with reticulate::use_condaenv(..., force=TRUE)
#>     If you prefer to let us automatically install a conda environment with 'mofapy2' installed using the 'basilisk' package, please use the argument 'use_basilisk = TRUE'
#> Error in run_mofa(MOFAObject.untrained, use_basilisk = FALSE, outfile = outfile,     save_data = TRUE): mofapy2_0.7.0 is not detected in the specified python binary, see reticulate::py_config(). Consider setting use_basilisk = TRUE to create a python environment with basilisk (https://bioconductor.org/packages/release/bioc/html/basilisk.html)

# Integration using MixOmics
mixObj <- prepareForIntegration(MAE,
                                omicsNames = c("protetest", "metatest"),
                                variableLists = rownames(MAE),
                                method = "mixOmics")
#> Warning: 'experiments' dropped; see 'drops()'
#> harmonizing input:
#>   removing 27 sampleMap rows not in names(experiments)
MAEtest <- runOmicsIntegration(MAEtest, preparedObject = mixObj, 
                               method = "mixOmics")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'runOmicsIntegration': object 'MAEtest' not found

# Access mixOmics results:
getMixOmics(MAEtest, response = "temperature")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getMixOmics': object 'MAEtest' not found
getMixOmicsSettings(MAEtest)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getMixOmicsSettings': object 'MAEtest' not found
mixOmics::plotIndiv(getMixOmics(MAEtest, response = "imbibition"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getMixOmics': object 'MAEtest' not found

# Access MOFA2 results:
getMOFA(MAEtest)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getMOFA': object 'MAEtest' not found
getMOFASettings(MAEtest)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getMOFASettings': object 'MAEtest' not found
MOFA2::plot_variance_explained(getMOFA(MAEtest))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'getMOFA': object 'MAEtest' not found