FlomicsMultiAssay.constructor Constructor for the class MultiAssayExperiment-class

This function initializes an object of class MultiAssayExperiment-class from a list of omics data and an object of class ExpDesign-class.

This function initializes an object of class MultiAssayExperiment-class from a list of omics data and an object of class ExpDesign-class.

FlomicsMultiAssay.constructor(inputs, Design, projectName)

Arguments

inputs
A named list of omic dataset. Names must refer to the name of the omic dataset. An omics dataset must be itself a list of three objects:

data:matrix of omic data

meta:an optional quality check data

omicType:Type of omic data type "None", "RNAseq", "proteomics" or "Metabolomics".

Design
An object of class ExpDesign-class

projectName
Project name

Value

An object of class MultiAssayExperiment-class

An object of class MultiAssayExperiment-class

Examples

Design.File <- read.table(file= paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/experimental_design.txt",sep=""),header = TRUE,row.names = 1, sep = "\t")

# Define the type of each factor
Design.Factors.Type <- c("Bio","Bio","batch")

# Define the reference modality for each factor
Design.Factors.Ref <- c("WT","control","rep1")

# Initialize an object of class ExpDesign
Design.obj <- ExpDesign.constructor(ExpDesign = Design.File, projectName = "Design.Name", refList = Design.Factors.Ref,
 typeList = Design.Factors.Type)
#> Error in ExpDesign.constructor(ExpDesign = Design.File, projectName = "Design.Name",     refList = Design.Factors.Ref, typeList = Design.Factors.Type) argument inutilisé (projectName = "Design.Name")
Design.Factors.Name <- names(Design.File)
Design.formulae <- GetModelFormulae(Factors.Name = Design.Factors.Name,Factors.Type=Design.Factors.Type)
Design.formulae[[1]]
#> ~replicat + genotype + treatment + genotype:treatment
#> <environment: 0x7f980f660f88>
Design.obj <- getExpressionContrast(object = Design.obj, model.formula = names(Design.formulae[1]))
#> Error in h(simpleError(msg, call)) erreur d'ï¿½valuation de l'argument 'object' lors de la sï¿½lection d'une mï¿½thode pour la fonction 'getExpressionContrast' : objet 'Design.obj' introuvable
Design.contrastList <- lapply(Design.obj@Contrasts.List, function(x) {
return(x[1:2]$contrast)
})
#> Error in lapply(Design.obj@Contrasts.List, function(x) {    return(x[1:2]$contrast)}) objet 'Design.obj' introuvable
Design.obj <- getContrastMatrix(object = Design.obj, contrastList = unlist(Design.contrastList))
#> Error in h(simpleError(msg, call)) erreur d'ï¿½valuation de l'argument 'object' lors de la sï¿½lection d'une mï¿½thode pour la fonction 'getContrastMatrix' : objet 'Design.obj' introuvable

 # Create a list of datasets
ListofData <- list("RNAseq1"=list("dataFile"=paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/rnaseq_gene_counts.txt",sep=""),
"qcFile"=paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/rnaseq_bioinfo_QC.txt",sep=""), "omicType"="RNAseq"))
FlomicsMultiAssay.constructor(inputs = ListofData, Design=Design.obj)
#> Error in array(x, c(length(x), 1L), if (!is.null(names(x))) list(names(x),     NULL) else NULL) 'data' doit être de type vecteur, il était 'NULL'

Design.File <- read.table(file= paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/experimental_design.txt",sep=""),header = TRUE,row.names = 1, sep = "\t")

# Define the type of each factor
Design.Factors.Type <- c("Bio","Bio","batch")

# Define the reference modality for each factor
Design.Factors.Ref <- c("WT","control","rep1")

# Initialize an object of class ExpDesign
Design.obj <- ExpDesign.constructor(ExpDesign = Design.File, projectName = "Design.Name", refList = Design.Factors.Ref,
 typeList = Design.Factors.Type)
#> Error in ExpDesign.constructor(ExpDesign = Design.File, projectName = "Design.Name",     refList = Design.Factors.Ref, typeList = Design.Factors.Type) argument inutilisé (projectName = "Design.Name")
Design.Factors.Name <- names(Design.File)
Design.formulae <- GetModelFormulae(Factors.Name = Design.Factors.Name,Factors.Type=Design.Factors.Type)
Design.formulae[[1]]
#> ~replicat + genotype + treatment + genotype:treatment
#> <environment: 0x7f98135ed140>
Design.obj <- getExpressionContrast(object = Design.obj, model.formula = names(Design.formulae[1]))
#> Error in h(simpleError(msg, call)) erreur d'ï¿½valuation de l'argument 'object' lors de la sï¿½lection d'une mï¿½thode pour la fonction 'getExpressionContrast' : objet 'Design.obj' introuvable
Design.contrastList <- lapply(Design.obj@Contrasts.List, function(x) {
return(x[1:2]$contrast)
})
#> Error in lapply(Design.obj@Contrasts.List, function(x) {    return(x[1:2]$contrast)}) objet 'Design.obj' introuvable
Design.obj <- getContrastMatrix(object = Design.obj, contrastList = unlist(Design.contrastList))
#> Error in h(simpleError(msg, call)) erreur d'ï¿½valuation de l'argument 'object' lors de la sï¿½lection d'une mï¿½thode pour la fonction 'getContrastMatrix' : objet 'Design.obj' introuvable

 # Create a list of datasets
ListofData <- list("RNAseq1"=list("dataFile"=paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/rnaseq_gene_counts.txt",sep=""),
"qcFile"=paste(path.package("RFLOMICS"),"/ExamplesFiles/TP/rnaseq_bioinfo_QC.txt",sep=""), "omicType"="RNAseq"))
FlomicsMultiAssay.constructor(inputs = ListofData, Design=Design.obj)
#> Error in array(x, c(length(x), 1L), if (!is.null(names(x))) list(names(x),     NULL) else NULL) 'data' doit être de type vecteur, il était 'NULL'